moved ErrorPage to app folder
