--- packages/component-elife-ui/client/index.js
@@ -1,1 +1,0 @@
-export { default as ErrorPage } from './pages/Error'
